Move node functionalities into platform_node
